+++ core/connectors/postgres/chat_memory_repo.js
@@ -104,3 +104,4 @@ class ChatMemoryRepo {
 module.exports = { ChatMemoryRepo };
 
 
+