+++ core/connectors/postgres/memory_suggestions_repo.js
@@ -89,3 +89,4 @@ class MemorySuggestionsRepo {
 module.exports = { MemorySuggestionsRepo, computeCandidateHash };
 
 
+