+++ core/connectors/postgres/work_context_repo.js
@@ -45,3 +45,4 @@ class WorkContextRepo {
 module.exports = { WorkContextRepo };
 
 
+